requirements.txt
gradio
numpy
openai-whisper
plotly
torch
transformers
